Add all router routes and functionality. -Refactor some comment endpoints that use the comments to table, to be located in articles controllers and models
