+++ app/controllers/articles.controllers.js
@@ -4,6 +4,8 @@ const {
   selectArticleByArticleId,
   selectAllArticles,
   updateVotesOnArticle,
+  selectCommentsByArticleId,
+  insertCommentToArticle,
 } = require("../models/articles.models");
 
 const getArticleByArticleId = (req, res, next) => {
@@ -116,4 +118,37 @@ const patchVotesOnArticle = (req, res, next) => {
     .catch(next);
 };
 
-module.exports = { getArticleByArticleId, getAllArticles, patchVotesOnArticle };
+const postCommentToArticle = (req, res, next) => {
+  const { article_id } = req.params;
+  const { username, body } = req.body;
+  const pendingSelectArticleById = selectArticleByArticleId(article_id);
+  const pendingInsertCommentsById = insertCommentToArticle(
+    article_id,
+    username,
+    body
+  );
+  Promise.all([pendingInsertCommentsById, pendingSelectArticleById])
+    .then(([newComment]) => {
+      res.status(201).send({ newComment });
+    })
+    .catch(next);
+};
+
+const getCommentsByArticleId = (req, res, next) => {
+  const { article_id } = req.params;
+  const pendingArticleById = selectArticleByArticleId(article_id);
+  const pendingSelectCommentsById = selectCommentsByArticleId(article_id);
+  Promise.all([pendingSelectCommentsById, pendingArticleById]) //result promise first, validity check promise second
+    .then(([comments]) => {
+      res.status(200).send({ comments });
+    })
+    .catch(next);
+};
+
+module.exports = {
+  getArticleByArticleId,
+  getAllArticles,
+  patchVotesOnArticle,
+  postCommentToArticle,
+  getCommentsByArticleId,
+};
