+++ app/controllers/users.controllers.js
@@ -8,4 +8,4 @@ const getAllUsers = (req, res, next) => {
     .catch(next);
 };
 
-module.exports = { getAllUsers };
+module.exports = getAllUsers;
