+++ app/models/articles.models.js
@@ -46,8 +46,40 @@ const updateVotesOnArticle = (voteIncrement, articleId) => {
     });
 };
 
+const selectCommentsByArticleId = (articleId) => {
+  return db
+    .query(
+      `SELECT * FROM comments 
+        WHERE article_id = $1
+        ORDER BY created_at ASC`,
+      [articleId]
+    )
+    .then(({ rows }) => {
+      return rows;
+    });
+};
+
+const insertCommentToArticle = (articleId, username, body) => {
+  if (!username || !body) {
+    return Promise.reject({ status: 400, msg: "Invalid request!" });
+  } else {
+    return db
+      .query(
+        `INSERT INTO comments (article_id, author, body) 
+        VALUES ($1, $2, $3)
+        RETURNING *`,
+        [articleId, username, body]
+      )
+      .then(({ rows }) => {
+        return rows[0];
+      });
+  }
+};
+
 module.exports = {
   selectArticleByArticleId,
   selectAllArticles,
   updateVotesOnArticle,
+  selectCommentsByArticleId,
+  insertCommentToArticle,
 };
