+++ app/routes/articles-router.js
@@ -4,8 +4,14 @@ const {
   getArticleByArticleId,
   getAllArticles,
   patchVotesOnArticle,
+  getCommentsByArticleId,
+  postCommentToArticle,
 } = require("../controllers/articles.controllers");
 
 articlesRouter.get("/", getAllArticles);
+articlesRouter.get("/:article_id", getArticleByArticleId);
+articlesRouter.patch("/:article_id", patchVotesOnArticle);
+articlesRouter.get("/:article_id/comments", getCommentsByArticleId);
+articlesRouter.post("/:article_id/comments", postCommentToArticle);
 
 module.exports = articlesRouter;
