+++ app/routes/topics-router.js
@@ -1,0 +1,7 @@
+const topicsRouter = require("express").Router();
+
+const getAllTopics = require("../controllers/topics.controllers");
+
+topicsRouter.get("/", getAllTopics);
+
+module.exports = topicsRouter;
