+++ app/routes/users-router.js
@@ -1,0 +1,7 @@
+const usersRouter = require("express").Router();
+
+const getAllUsers = require("../controllers/users.controllers");
+
+usersRouter.get("/", getAllUsers);
+
+module.exports = usersRouter;
